Break engine into workable pieces
